code_compare_ai/requirements.txt
streamlit>=1.40.0
requests>=2.32.0
python-dotenv>=1.0.1
openai>=1.40.0
google-genai>=0.3.0
